common/mcp_armor/requirements.txt
# MCP Content Filter Dependencies

# Core dependencies
pyyaml>=6.0